Refactor wormhole server
use one flat Options struct for both the controller and the tunnel.
They are not meant to be run independently and share some things (e.g. logger)

Also consolidated the server in a single package because
